test(Line): 기존 작성된 인수테스트 폐기하고 새로 작성된 인수테스트로 교체
